keep: warningレベルを上げた際に発生するwarningの一部を修正 以下のwarningを修正しました 　warning C4244: 'xx' : 'yyyy' から 'zzzz' への変換です。データが失われる可能性があります。 　warning C4310: キャストによって定数値が切り捨てられました。 　　(ANSI構成での1箇所対応しない箇所が有ります) 　warning C4701: 初期化されていない可能性のあるローカル変数 'xxxx' が使用されます 　warning C4702: 制御が渡らないコードです。 VS2010Expressで評価しました Patchunicode#372
git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk2@2745 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/func/CFuncKeyWnd.cpp
@@ -324,7 +324,7 @@ LRESULT CFuncKeyWnd::OnTimer( HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam
 		for( i = 0; i < _countof( m_szFuncNameArr ); ++i ){
 			// 2007.02.22 ryoji CKeyBind::GetFuncCode()‚ðŽg‚¤
 			EFunctionCode	nFuncCode = CKeyBind::GetFuncCode(
-					(((VK_F1 + i) | ((WORD)((BYTE)(nIdx))) << 8)),
+					(WORD)(((VK_F1 + i) | ((WORD)((BYTE)(nIdx))) << 8)),
 					m_pShareData->m_Common.m_sKeyBind.m_nKeyNameArrNum,
 					m_pShareData->m_Common.m_sKeyBind.m_pKeyNameArr
 			);
